+++ comet_pqc/dashboard.py
@@ -625,6 +625,7 @@ class Dashboard(ui.Row, ProcessMixin, SettingsMixin, ResourceMixin):
             for sequence in self.sequence_combobox:
                 if sequence.id == current_sequence_id:
                     self.sequence_combobox.current = sequence
+                    self.on_load_sequence_tree(sequence.id)
                     break
 
     def sample_name(self):
@@ -634,6 +635,19 @@ class Dashboard(ui.Row, ProcessMixin, SettingsMixin, ResourceMixin):
     def sample_type(self):
         """Return sample type."""
         return self.sample_type_text.value.strip()
+
+    def table_position(self):
+        """Return table position as tuple. If table not available return
+        (0, 0, 0).
+        """
+        if self.use_table:
+            try:
+                with self.resources.get("table") as table_resource:
+                    table = Venus1(table_resource)
+                    return table.pos
+            except:
+                pass
+        return (0, 0, 0)
 
     def use_environment(self):
         """Return True if environment box enabled."""
@@ -780,6 +794,7 @@ class Dashboard(ui.Row, ProcessMixin, SettingsMixin, ResourceMixin):
         sequence = self.processes.get("sequence")
         sequence.set("sample_name", self.sample_name())
         sequence.set("sample_type", self.sample_type())
+        sequence.set("table_position", self.table_position())
         sequence.set("operator", self.operator())
         sequence.set("output_dir", self.output_dir())
         sequence.set("write_logfiles", self.settings.get("write_logfiles", True))
@@ -858,6 +873,7 @@ class Dashboard(ui.Row, ProcessMixin, SettingsMixin, ResourceMixin):
         # Set process parameters
         measure.set("sample_name", self.sample_name())
         measure.set("sample_type", self.sample_type())
+        measure.set("table_position", self.table_position())
         measure.set("operator", self.operator())
         measure.set("output_dir", self.output_dir())
         measure.set("write_logfiles", self.settings.get("write_logfiles", True))